// sdks/node/v1/src/models/ContactMethod.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Interactions API
 * Interactions API documentation
 *
 * The version of the OpenAPI document: v1
 * Contact: api@demexchange.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ContactMethod = {
    Unknown: 'unknown',
    Mail: 'mail',
    Letter: 'letter',
    DigitalAd: 'digital_ad',
    Email: 'email',
    Text: 'text',
    TextBroadcast: 'text_broadcast',
    RoboCall: 'robo_call',
    DialerCall: 'dialer_call',
    PhoneCall: 'phone_call',
    DoorKnock: 'door_knock',
    Event: 'Event',
    HotSpot: 'hot_spot',
    OneOnOne: 'one_on_one',
    WebInteraction: 'web_interaction'
} as const;
export type ContactMethod = typeof ContactMethod[keyof typeof ContactMethod];


export function instanceOfContactMethod(value: any): boolean {
    for (const key in ContactMethod) {
        if (Object.prototype.hasOwnProperty.call(ContactMethod, key)) {
            if (ContactMethod[key as keyof typeof ContactMethod] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ContactMethodFromJSON(json: any): ContactMethod {
    return ContactMethodFromJSONTyped(json, false);
}

export function ContactMethodFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContactMethod {
    return json as ContactMethod;
}

export function ContactMethodToJSON(value?: ContactMethod | null): any {
    return value as any;
}

export function ContactMethodToJSONTyped(value: any, ignoreDiscriminator: boolean): ContactMethod {
    return value as ContactMethod;
}