// sdks/node/v1/src/models/Outcome.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Interactions API
 * Interactions API documentation
 *
 * The version of the OpenAPI document: v1
 * Contact: api@demexchange.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Outcome = {
    Unknown: 'unknown',
    ClickedLink: 'clicked_link',
    EmailOpened: 'email_opened',
    SubmittedForm: 'submitted_form',
    SuccessfulContact: 'successful_contact',
    Deceased: 'deceased',
    DeliverabilityError: 'deliverability_error',
    PermanentlyUndeliverable: 'permanently_undeliverable',
    Inaccessible: 'inaccessible',
    LanguageBarrier: 'language_barrier',
    Moved: 'moved',
    NoAnswer: 'no_answer',
    WrongNumber: 'wrong_number',
    Hostile: 'hostile',
    MarkedSpam: 'marked_spam',
    Refused: 'refused',
    Delivered: 'delivered',
    LitDrop: 'lit_drop',
    LeftMessage: 'left_message',
    Other: 'other'
} as const;
export type Outcome = typeof Outcome[keyof typeof Outcome];


export function instanceOfOutcome(value: any): boolean {
    for (const key in Outcome) {
        if (Object.prototype.hasOwnProperty.call(Outcome, key)) {
            if (Outcome[key as keyof typeof Outcome] === value) {
                return true;
            }
        }
    }
    return false;
}

export function OutcomeFromJSON(json: any): Outcome {
    return OutcomeFromJSONTyped(json, false);
}

export function OutcomeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Outcome {
    return json as Outcome;
}

export function OutcomeToJSON(value?: Outcome | null): any {
    return value as any;
}

export function OutcomeToJSONTyped(value: any, ignoreDiscriminator: boolean): Outcome {
    return value as Outcome;
}