// sdks/node/v1/src/models/InteractionsDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Interactions API
 * Interactions API documentation
 *
 * The version of the OpenAPI document: v1
 * Contact: api@demexchange.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { InteractionDto } from './InteractionDto';
import {
    InteractionDtoFromJSON,
    InteractionDtoFromJSONTyped,
    InteractionDtoToJSON,
    InteractionDtoToJSONTyped,
} from './InteractionDto';

/**
 * 
 * @export
 * @interface InteractionsDto
 */
export interface InteractionsDto {
    /**
     * 
     * @type {Array<InteractionDto>}
     * @memberof InteractionsDto
     */
    interactions?: Array<InteractionDto>;
}

/**
 * Check if a given object implements the InteractionsDto interface.
 */
export function instanceOfInteractionsDto(value: object): value is InteractionsDto {
    return true;
}

export function InteractionsDtoFromJSON(json: any): InteractionsDto {
    return InteractionsDtoFromJSONTyped(json, false);
}

export function InteractionsDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): InteractionsDto {
    if (json == null) {
        return json;
    }
    return {
        
        'interactions': json['interactions'] == null ? undefined : ((json['interactions'] as Array<any>).map(InteractionDtoFromJSON)),
    };
}

export function InteractionsDtoToJSON(json: any): InteractionsDto {
    return InteractionsDtoToJSONTyped(json, false);
}

export function InteractionsDtoToJSONTyped(value?: InteractionsDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'interactions': value['interactions'] == null ? undefined : ((value['interactions'] as Array<any>).map(InteractionDtoToJSON)),
    };
}